libbun: add support for libunwind.
Summary:
this diff adds the libunwind backend to libbun along with a
crashpad example utilizing it.

To enable the backend, libunwind must be available on the build system
and LIBUNWIND_ENABLED cmake variable needs to be set.

The major things introduced by this diff:
 - the libunwind backend, to be found in src/backend/libunwind,
  bun_libunwind.c specifically
 - a test file
 - CMake module files for searching for libunwind and crashpad
 - an example program using libbun + crashpad

Depends on D3451

Test Plan: gtest

Reviewers: sbahra, pkhuong, ahicks, rqurashi

Differential Revision: https://hack.in.backtrace.io/D3457